pyproject.toml
[tool.poetry]
name = "movie-recommendation"
version = "0.1.0"
description = "This is a movie recommendation system"
authors = ["MikeRoonane <mikeroonane2005@gmail.com>"]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "services" },
]

[tool.poetry.dependencies]
python = ">=3.12"
scikit-learn = "^1.4.1"  
scikit-surprise = "^1.1.3"      
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
notebook = "^7.2.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"